Refactor sources

See merge request tischer/platy-browser-tables!3

+++ scripts/files/copy_helper.py
@@ -1,7 +1,7 @@
 import os
-import glob
 import shutil
 from .xml_utils import copy_xml_with_newpath, get_h5_path_from_xml
+from .sources import get_image_names, get_segmentation_names, get_segmentations
 
 
 def copy_file(xml_in, xml_out):
@@ -9,17 +9,6 @@ def copy_file(xml_in, xml_out):
     xml_dir = os.path.split(xml_out)[0]
     h5path = os.path.relpath(h5path, start=xml_dir)
     copy_xml_with_newpath(xml_in, xml_out, h5path, path_type='relative')
-
-
-def copy_files_with_pattern(src_folder, dst_folder, pattern):
-    files = glob.glob(os.path.join(src_folder, pattern))
-    for ff in files:
-        ext = os.path.splitext(ff)[1]
-        if ext != '.xml':
-            continue
-        xml_in = ff
-        xml_out = os.path.join(dst_folder, os.path.split(ff)[1])
-        copy_file(xml_in, xml_out)
 
 
 # For now we put symlinks with relative paths, but I am not sure
@@ -45,6 +34,8 @@ def copy_segmentation(src_folder, dst_folder, name):
     # copy the segmentation xml
     name_with_ext = '%s.xml' % name
     xml_in = os.path.join(src_folder, 'segmentations', name_with_ext)
+    if not os.path.exists(xml_in):
+        raise RuntimeError("Could not find %s in the src folder %s" % (name, src_folder))
     xml_out = os.path.join(dst_folder, 'segmentations', name_with_ext)
     copy_file(xml_in, xml_out)
 
@@ -60,28 +51,51 @@ def copy_segmentation(src_folder, dst_folder, name):
 
 
 def copy_image_data(src_folder, dst_folder):
-    # copy sbem image data
-    sbem_prefix = 'sbem-6dpf-1-whole'
-    raw_name = '%s-raw.xml' % sbem_prefix
-    copy_file(os.path.join(src_folder, raw_name),
-              os.path.join(dst_folder, raw_name))
+    # get all image names that need to be copied
+    names = get_image_names()
 
-    # copy the prospr med image data
-    copy_files_with_pattern(src_folder, dst_folder, '*-MED*')
-
-    # copy the segmented prospr regions
-    copy_files_with_pattern(src_folder, dst_folder, 'prospr-6dpf-1-whole-segmented-*')
+    for name in names:
+        name += '.xml'
+        in_path = os.path.join(src_folder, 'images', name)
+        out_path = os.path.join(dst_folder, 'images', name)
+        if not os.path.exists(in_path):
+            raise RuntimeError("Could not find %s in the src folder %s" % (name, src_folder))
+        # copy the xml
+        copy_file(in_path, out_path)
 
 
 def copy_misc_data(src_folder, dst_folder):
     # copy the aux gene data
     prospr_prefix = 'prospr-6dpf-1-whole'
     aux_name = '%s_meds_all_genes.xml' % prospr_prefix
-    copy_file(os.path.join(src_folder, aux_name),
-              os.path.join(dst_folder, aux_name))
+    copy_file(os.path.join(src_folder, 'misc', aux_name),
+              os.path.join(dst_folder, 'misc', aux_name))
 
     # copy the bookmarks
-    bkmrk_in = os.path.join(src_folder, 'bookmarks.json')
+    bkmrk_in = os.path.join(src_folder, 'misc', 'bookmarks.json')
     if os.path.exists(bkmrk_in):
         shutil.copyfile(bkmrk_in,
-                        os.path.join(dst_folder, 'bookmarks.json'))
+                        os.path.join(dst_folder, 'misc', 'bookmarks.json'))
+
+
+def copy_segmentations(src_folder, dst_folder):
+    names = get_segmentation_names()
+    for name in names:
+        copy_segmentation(src_folder, dst_folder, name)
+
+
+def copy_all_tables(src_folder, dst_folder):
+    segmentations = get_segmentations()
+    for name, seg in segmentations.items():
+        has_table = seg.get('has_tables', False) or 'table_update_function' in seg
+        if not has_table:
+            continue
+        copy_tables(src_folder, dst_folder, name)
+
+
+def copy_release_folder(src_folder, dst_folder):
+    # copy static image and misc data
+    copy_image_data(src_folder, dst_folder)
+    copy_misc_data(src_folder, dst_folder)
+    copy_segmentations(src_folder, dst_folder)
+    copy_all_tables(src_folder, dst_folder)
